+++ Back/Nenufar/src/main/java/com/nenuphar/nenufar/Repositories/SubSkillRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface SubSkillRepository extends CrudRepository<SubSkill, Long> {
     @Query(value="SELECT * FROM sub_skill WHERE skill_id = ?1", nativeQuery = true)
-    List<SubSkill> getSubSkillsFromSkillID(int skill_id);
+    List<SubSkill> getSubSkillsFromSkillID(Long skill_id);
 }
