+++ Back/Nenufar/src/main/java/com/nenuphar/nenufar/Services/SubSkillService.java
@@ -24,7 +24,7 @@ public class SubSkillService {
         return subskill;
     }
 
-    public List<SubSkill> getSubSkillsFromSkillID(int skill_id)
+    public List<SubSkill> getSubSkillsFromSkillID(Long skill_id)
     {
         try
         {
